+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Terms of Delivery 
 - Terms of Payment
 - Way of Delivery
+- Company Information
 
 ### Fixed
 - Typoes with currency that shoud be currencies